requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
cffi>=1.15.0
cryptography>=41.0.0

